Fix HelicopterSlowDeathBehavior crash
Yes, this is technically a design/scripting issue, but in our case, we used this on naval units to have a fancy death effect. If a naval unit spawned on land, and died, the game would crash.

This nullptr check addresses the issue, and it's obvious that naval units are not meant to spawn on ground. This is fine.